:construction: WIP: update the nav button
